create addExpense.js
