+++ frontend/src/js/addExpense.js
@@ -1,0 +1,33 @@
+import '../css/style.css';
+import JwtService from './service/JwtService.js';
+import ConfigApp from './config.js';
+import RequestService from './service/RequestService.js';
+
+const config = new ConfigApp('prod');
+const jwtService = new JwtService();
+const requestService = new RequestService(config.getRestUrl());
+
+if (jwtService.checkExpire()) {
+    jwtService.logout();
+}
+
+//if ("serviceWorker" in navigator) {
+//    window.addEventListener("load", function () {
+//        navigator.serviceWorker
+//            .register("/serviceWorker.js")
+//            .then(res => console.log("service worker registered"))
+//            .catch(err => console.log("service worker not registered", err));
+//    });
+//}
+
+document.getElementById('logout-button').addEventListener('click', () => jwtService.logout());
+
+const addExpenseForm = document.getElementById('add-expense-form');
+if (addExpenseForm) {
+    addExpenseForm.addEventListener('submit', function (e) {
+        requestService.submitExpenseForm(e, this).then(() => {
+            alert('Expense added!');
+            window.location.href = '/week.html';
+        }).catch(() => alert('Cannot add expense. Please try again'));
+    });
+}
